scripts/image-bot/requirements.txt
discord.py==1.5.1
google-api-python-client==1.30.0
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
requests==2.25.1
